Move shared functions to Item and add download
