+++ features/gdrive_abstractions.feature
@@ -34,3 +34,20 @@ Feature: Box File and Folder abstractions
         When I call the function: "get_parent" on saved object: "root"
         Then the result is null
 
+    Scenario: Download file
+        When I get a file with id: "1RpxzKzzjDH57n3Vs-RroFf-U_dT_gPtW" and save it as: "to_download"
+        And I call the function: "download_to" on saved object: "to_download" with parameters: "test.txt"
+        Then the local file "test.txt" exists
+        And delete the local file "test.txt"
+
+    Scenario: Get name of file
+        When I get a file with id: "1RpxzKzzjDH57n3Vs-RroFf-U_dT_gPtW" and save it as: "other"
+        And I call the function: "get_name" on saved object: "other"
+        Then the result should equal: "Lorem.txt"
+
+    Scenario: Get parent of file
+        When I get a file with id: "1RpxzKzzjDH57n3Vs-RroFf-U_dT_gPtW" and save it as: "other"
+        And I call the function: "get_parent" on saved object: "other"
+        And I save the result as: "parent"
+        And I call the function: "get_name" on saved object: "parent"
+        Then the result should equal: "Text Files"
