+++ lib/integrations/Box/abstractions/file.js
@@ -1,4 +1,4 @@
-const Folder = require("./folder");
+const Item = require("./item");
 
 /**
  * File class
@@ -11,5 +11,25 @@ function File(id, connection) {
   this.type = "file";
   this.connection = connection;
 }
+
+File.prototype = Object.create(Item.prototype);
+File.prototype.constructor = File;
+
+/**
+ * Downloads this file to specified location
+ * @param {String} path - Path to destination.
+ *
+ * @return {Promise}
+ */
+File.prototype.download_to = function(path) {
+  return new Promise(async (resolve, reject) => {
+    try {
+      await this.connection.download_file(this.id, path);
+      resolve();
+    } catch (err) {
+      reject(err);
+    }
+  });
+};
 
 module.exports = File;
