+++ lib/integrations/Box/abstractions/folder.js
@@ -1,4 +1,4 @@
-const File = require("./file");
+const Item = require("./item");
 
 /**
  * Folder class
@@ -15,12 +15,16 @@ function Folder(id, connection) {
   this.connection = connection;
 }
 
+Folder.prototype = Object.create(Item.prototype);
+Folder.prototype.constructor = Folder;
+
 /**
  * List the files in this folder.
  *
  * @return {Promise}
  */
 Folder.prototype.list_files = function() {
+  const File = require("./file");
   return new Promise(async (resolve, reject) => {
     try {
       let res = (await this.connection.list_all_files(this.id)).entries;
@@ -39,69 +43,6 @@ Folder.prototype.list_files = function() {
     } catch (err) {
       reject(err);
     }
-  });
-};
-
-/**
- * Retrieves info for this folder and adds that info to the object.
- *
- * @return {Promise}
- */
-Folder.prototype.retrieve_info = function() {
-  return new Promise(async (resolve, reject) => {
-    try {
-      let result = await this.connection.get_folder_info(this.id);
-      this.name = result.name;
-      if (result.parent) {
-        this.parent = new Folder(result.parent.id, this.connection);
-        this.parent.name = result.parent.name;
-      } else {
-        this.parent = null;
-      }
-      this.size = result.size;
-      this.trashed = result.trashed_at != null;
-      resolve();
-    } catch (err) {
-      reject(err);
-    }
-  });
-};
-
-/**
- * Get the name of this folder.
- *
- * @return {Promise}
- */
-Folder.prototype.get_name = function() {
-  return new Promise(async (resolve, reject) => {
-    if (!this.name) {
-      try {
-        await this.retrieve_info();
-      } catch (err) {
-        reject(err);
-        return;
-      }
-    }
-    resolve(this.name);
-  });
-};
-
-/**
- * Get the parent folder of this folder or null if there isn't one.
- *
- * @return {Folder}
- */
-Folder.prototype.get_parent = function() {
-  return new Promise(async (resolve, reject) => {
-    if (this.parent == undefined) {
-      try {
-        await this.retrieve_info();
-      } catch (err) {
-        reject(err);
-        return;
-      }
-    }
-    resolve(this.parent);
   });
 };
