+++ lib/integrations/GoogleDrive/abstractions/__tests__/file_spec.js
@@ -2,9 +2,9 @@ const File = require("../file");
 
 describe("file", () => {
   describe("constructor", () => {
-    const expected_id = "1234";
-    const expected_type = "file";
-    const expected_connection = "a_connection";
+    const expected_id = "1234",
+      expected_connection = "the_connection",
+      expected_type = "file";
     let actual;
     beforeEach(() => {
       actual = new File(expected_id, expected_connection);
@@ -14,12 +14,12 @@ describe("file", () => {
       expect(actual.id).toEqual(expected_id);
     });
 
-    it("sets the type", () => {
-      expect(actual.type).toEqual(expected_type);
-    });
-
     it("sets the connection", () => {
       expect(actual.connection).toEqual(expected_connection);
+    });
+
+    it("sets the type", () => {
+      expect(actual.type).toEqual(expected_type);
     });
   });
